// src/app/modules/business-common/src/lib/components/form-summary/models/form-summary-buttons-config.model.ts
import { ButtonConfig } from '@dcx/ui/libs';

export interface FormSummaryButtonsConfig {
  addButton?: ButtonConfig;
  cancelButton?: ButtonConfig;
  editButton?: ButtonConfig;
  saveButton?: ButtonConfig;
  hideAddButton?: boolean;
  hideCancelButton?: boolean;
  hideEditButton?: boolean;
  hideSaveButton?: boolean;
}